supernova/scsynth: Buffer reading should send /b_info
Buffer reading issues two asynchronous commands: first it sends the buffer
filling message, then a /b_query to obtain information from the server.

this means that after waiting for Server.sync, the Buffer object is not
initialized. we avoid this server/language ping-pong by automatically
sending /b_info to avoid that the language has to explicitly send a
/b_query

Signed-off-by: Tim Blechmann <tim@klingt.org>

Conflicts:
	SCClassLibrary/Common/Control/Buffer.sc